backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.10.5
click==8.3.0
cryptography==42.0.5
distro==1.9.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx>=0.26.0
idna==3.11
jiter==0.11.0
openai==2.3.0
pydantic==2.12.1
pydantic_core==2.41.3
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
supabase==2.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
websockets>=13.0

# Advanced Hallucination Detection Dependencies
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
hdbscan==0.8.33
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.3